requirements-rag.txt
google-genai
qdrant-client
rank-bm25
typer
rich
pydantic
python-dotenv
